requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
joblib>=1.0.0
# GPU dependencies (optional)
cudf-cu11>=21.10.0
cuml-cu11>=21.10.0
cupy-cuda11x>=9.0.0 